patricia_tree::set
Struct GenericPatriciaSet

pub struct GenericPatriciaSet<T> { /* private fields */ }

Patricia tree based set.

Implementations

impl<T> GenericPatriciaSet<T>

pub fn new() -> Self

Makes a new empty GenericPatriciaSet instance.

Examples

use patricia_tree::PatriciaSet;

let set = PatriciaSet::new();
assert!(set.is_empty());

pub fn len(&self) -> usize

Returns the number of elements in this set.

Examples

use patricia_tree::PatriciaSet;

let mut set = PatriciaSet::new();
set.insert("foo");
set.insert("bar");
assert_eq!(set.len(), 2);

pub fn is_empty(&self) -> bool

Returns true if this set contains no elements.

Examples

use patricia_tree::PatriciaSet;

let mut set = PatriciaSet::new();
assert!(set.is_empty());

set.insert("foo");
assert!(!set.is_empty());

set.clear();
assert!(set.is_empty());

pub fn clear(&mut self)

Clears this set, removing all values.

Examples

use patricia_tree::PatriciaSet;

let mut set = PatriciaSet::new();
set.insert("foo");
set.clear();
assert!(set.is_empty());

impl<T: Bytes> GenericPatriciaSet<T>

pub fn contains<U: AsRef<T::Borrowed>>(&self, value: U) -> bool

Returns true if this set contains a value.

Examples

use patricia_tree::PatriciaSet;

let mut set = PatriciaSet::new();
set.insert("foo");
assert!(set.contains("foo"));
assert!(!set.contains("bar"));

pub fn get_longest_common_prefix<'a, U>( &self, value: &'a U, ) -> Option<&'a T::Borrowed>
where U: ?Sized + AsRef<T::Borrowed>,

Finds the longest common prefix of value and the elements in this set.

Examples

use patricia_tree::PatriciaSet;

let mut set = PatriciaSet::new();

set.insert("foo");
set.insert("foobar");
assert_eq!(set.get_longest_common_prefix("fo"), None);
assert_eq!(set.get_longest_common_prefix("foo"), Some("foo".as_bytes()));
assert_eq!(set.get_longest_common_prefix("fooba"), Some("foo".as_bytes()));
assert_eq!(set.get_longest_common_prefix("foobar"), Some("foobar".as_bytes()));
assert_eq!(set.get_longest_common_prefix("foobarbaz"), Some("foobar".as_bytes()));

pub fn longest_common_prefix_len<U>(&self, value: &U) -> usize
where U: ?Sized + AsRef<T::Borrowed>,

Returns the longest common prefix length of value and the elements in this set.

Unlike get_longest_common_prefix(), this method does not check if there is a element that matches the prefix in this set.

Examples

use patricia_tree::PatriciaSet;

let mut set = PatriciaSet::new();
set.insert("foo");
set.insert("foobar");
assert_eq!(set.longest_common_prefix_len("fo"), 2);
assert_eq!(set.longest_common_prefix_len("foo"), 3);
assert_eq!(set.longest_common_prefix_len("fooba"), 5);
assert_eq!(set.longest_common_prefix_len("foobar"), 6);
assert_eq!(set.longest_common_prefix_len("foobarbaz"), 6);
assert_eq!(set.longest_common_prefix_len("foba"), 2);

pub fn insert<U: AsRef<T::Borrowed>>(&mut self, value: U) -> bool

Adds a value to this set.

If the set did not have this value present, true is returned. If the set did have this value present, false is returned, and the entry is not updated.

Examples

use patricia_tree::PatriciaSet;

let mut set = PatriciaSet::new();
assert!(set.insert("foo"));
assert!(!set.insert("foo"));
assert_eq!(set.len(), 1);

pub fn remove<U: AsRef<T::Borrowed>>(&mut self, value: U) -> bool

Removes a value from the set. Returns true is the value was present in this set.

Examples

use patricia_tree::PatriciaSet;

let mut set = PatriciaSet::new();
set.insert("foo");
assert_eq!(set.remove("foo"), true);
assert_eq!(set.remove("foo"), false);

pub fn split_by_prefix<U: AsRef<T::Borrowed>>(&mut self, prefix: U) -> Self

Splits the set into two at the given prefix.

The returned set contains all the entries that prefixed by prefix.

Examples

use patricia_tree::PatriciaSet;

let mut a = PatriciaSet::new();
a.insert("rust");
a.insert("ruby");
a.insert("python");
a.insert("erlang");

let b = a.split_by_prefix("ru");

assert_eq!(a.iter().collect::<Vec<_>>(), [b"erlang", b"python"]);
assert_eq!(b.iter().collect::<Vec<_>>(), [b"ruby", b"rust"]);

pub fn iter(&self) -> Iter<'_, T>

Gets an iterator over the contents of this set, in sorted order.

Examples

use patricia_tree::PatriciaSet;

let mut set = PatriciaSet::new();
set.insert("foo");
set.insert("bar");
set.insert("baz");

assert_eq!(set.iter().collect::<Vec<_>>(), [Vec::from("bar"), "baz".into(), "foo".into()]);

impl<T: Bytes> GenericPatriciaSet<T>

pub fn iter_prefix<'a, 'b>( &'a self, prefix: &'b T::Borrowed, ) -> impl 'a + Iterator<Item = T>
where 'b: 'a,

Gets an iterator over the contents having the given prefix of this set, in sorted order.

Examples

use patricia_tree::PatriciaSet;

let mut set = PatriciaSet::new();
set.insert("foo");
set.insert("bar");
set.insert("baz");

assert_eq!(set.iter_prefix(b"ba").collect::<Vec<_>>(), [Vec::from("bar"), "baz".into()]);