patricia_tree
Trait BorrowedBytes

pub trait BorrowedBytes {
    // Required methods
    fn as_bytes(&self) -> &[u8] ;
    fn is_valid_bytes(bytes: &[u8]) -> bool;
    fn from_bytes(bytes: &[u8]) -> &Self;
    fn strip_common_prefix(&self, bytes: &[u8]) -> &Self;
    fn cmp_first_item(&self, bytes: &[u8]) -> Ordering;
    fn strip_n_prefix(&self, n: usize) -> &Self;

    // Provided methods
    fn strip_common_prefix_and_len(&self, bytes: &[u8]) -> (&Self, usize) { ... }
    fn is_empty(&self) -> bool { ... }
}

Borrowed type of Bytes.

Required Methods

fn as_bytes(&self) -> &[u8]

Returns the byte representation of this instance.

fn is_valid_bytes(bytes: &[u8]) -> bool

Returns true if the given bytes is a valid representation of this type, otherwise false.

fn from_bytes(bytes: &[u8]) -> &Self

Converts the given bytes to an instance of this type.

Caller can assume that is_valid_bytes(bytes) is true.

fn strip_common_prefix(&self, bytes: &[u8]) -> &Self

Returns a suffix of this instance not containing the common prefix with the given bytes.

fn cmp_first_item(&self, bytes: &[u8]) -> Ordering

Compares the first item of this instance with the first item represented in the the given bytes.

fn strip_n_prefix(&self, n: usize) -> &Self

Returns a suffix of this instance not containing the first n bytes.

Provided Methods

fn strip_common_prefix_and_len(&self, bytes: &[u8]) -> (&Self, usize)

Same as [strip_common_prefix()], but also returns the length of the common prefix.

fn is_empty(&self) -> bool

Returns true if this instance is empty, otherwise false.

Object Safety

This trait is not object safe.