patricia_tree
Trait Bytes

pub trait Bytes {
    type Borrowed: ?Sized + BorrowedBytes + ToOwned<Owned = Self>;
}

This trait represents a bytes type that can be used as the key type of patricia trees.

Required Associated Types

type Borrowed: ?Sized + BorrowedBytes + ToOwned<Owned = Self>

Borrowed type of this type.